safe-ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "safe-ai-governor"
version = "1.0.0"
description = "Cryptographically-gated AI governor with policy VM and sandboxed execution"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "SAFE-AI Team", email = "team@safe-ai.example"},
]
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "PyNaCl==1.5.0",
    "cryptography==42.0.0",
    "PyJWT==2.8.0",
    "click==8.1.7",
    "pyyaml==6.0.1",
    "python-multipart==0.0.6",
    "prometheus-client==0.19.0",
    "python-json-logger==2.0.7",
    "httpx==0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "mypy==1.8.0",
    "ruff==0.1.14",
    "types-PyYAML==6.0.12.12",
]

[project.scripts]
safeai = "app.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "B", "ANN", "S", "C90"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ANN"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
asyncio_mode = "auto"
